change back to original for Ben packages.
I put the logger function in another file has it has nothing to do in base.
