+++ geoapps_utils/base.py
@@ -7,10 +7,8 @@
 #  (see LICENSE file at the root of this source code package).                      '
 #                                                                                   '
 # '''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
-
 from __future__ import annotations
 
-import logging
 import sys
 import tempfile
 import warnings
@@ -29,40 +27,10 @@ from typing_extensions import Self
 from geoapps_utils.driver.params import BaseParams
 from geoapps_utils.utils.formatters import recursive_flatten
 from geoapps_utils.utils.importing import GeoAppsError
+from geoapps_utils.utils.logger import get_logger
 
 
-def get_logger(name: str | None = None, timestamp: bool = False) -> logging.Logger:
-    """
-    Get a logger with a timestamped stream and specified log level.
-
-    :param name: Name of the logger.
-    :param timestamp: Whether to include a timestamp in the log format.
-    """
-    log = logging.getLogger(name)
-
-    if log.handlers:
-        stream_handler = log.handlers[0]
-    else:
-        stream_handler = logging.StreamHandler()
-
-    # Set the format for the logger
-    formatting = "%(levelname)s"
-
-    if name:
-        formatting += " [%(name)s]"
-
-    if timestamp:
-        formatting += " %(asctime)s"
-
-    formatter = logging.Formatter(formatting + " %(message)s")
-    stream_handler.setFormatter(formatter)
-    log.addHandler(stream_handler)
-
-    return log
-
-
-logging.basicConfig(level=logging.INFO, format="%(levelname)s %(message)s")
-logger = logging.getLogger(__name__)
+logger = get_logger(name=__name__)
 
 
 class Driver(ABC):
